requirements.txt
aiohttp
async-timeout
attrs
beautifulsoup4
cairocffi
certifi
cffi
chardet
charset-normalizer
chess.com
configparser
cssselect2
defusedxml
discord-webhook
discord.py
emoji
idna
lxml
multidict
nekos.py
Pillow
pornhub-api
pornhubapi
psycopg2
pycparser
pydantic
requests
soupsieve
tinycss2
typing-extensions
urllib3
webencodings
yarl
chess
python-urbandict
jishaku
git+https://github.com/Rapptz/discord-ext-menus
python-dotenv
asgiref
openai
revChatGPT
GoogleBard
EdgeGPT
selenium
undetected-chromedriver